Revert "Try using LiveComponents."
This reverts commit ec83dbf27c61ec9e8363a9071197e8d7c4b58fc1.
